Updated Makefile and install script.
This commit removes the legacy install process from the Makefile
and replaces it with the target that was called 'newinstall' in prior
commits.

It also tweaks the installer script to include a recompilation of the
scripts/ directory on the Pi post-install. This is a quick-and-dirty
way to get around the kernel distro's hard-coded HOSTCC behavior for
the scripts/ folder, and ensures that kernel modules will be
compilable directly on the Pi.